requirements.txt
numpy>=1.24.3
pandas>=1.5.3
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn>=1.2.2
beautifulsoup4>=4.12.2
jupyter>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
kaggle>=1.7.4
tqdm>=4.67.1
folium>=0.19.5
branca>=0.8.1
xyzservices>=2025.1.0
